requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
Django==4.2.11
djangorestframework==3.15.1
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.17.1
packaging==24.0
postgrest==0.16.3
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.2
supafunc==0.4.5
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
